Extract cinema option helpers
