+++ src/routes/movie/[id]/+page.server.ts
@@ -1,6 +1,6 @@
 import { error } from "@sveltejs/kit";
+import { get_cinema_options } from "$lib/cinemas";
 import { readMovies } from "$lib/movies";
-import { CAPITAL_REGION_CINEMAS } from "$lib/constants";
 import type { PageServerLoad } from "./$types";
 
 // The prerenderer can't find the /movie/[id] routes because it needs the entries() function to know which routes to prerender.
@@ -17,18 +17,7 @@ export const load: PageServerLoad = async ({ params }) => {
     error(404, "Movie not found");
   }
 
-  const all_cinemas = movies
-    .flatMap((movie) => Object.values(movie.showtimes_by_day).flatMap((day) => Object.keys(day)))
-    .filter((name, index, array) => array.indexOf(name) === index)
-    .sort();
-
-  const capital_region_cinemas = all_cinemas.filter((name) => (CAPITAL_REGION_CINEMAS as readonly string[]).includes(name));
-
-  const cinema_options = [
-    ["Öll kvikmyndahús", all_cinemas],
-    ["Höfuðborgarsvæðið", capital_region_cinemas],
-    ...all_cinemas.map((name) => [name, [name]] as const),
-  ] as const;
+  const cinema_options = get_cinema_options(movies);
 
   return {
     movie,
